### 2019-11-01 #1859 @whedon
[PRE REVIEW]: NeuroDiffEq: A Python package for solving differential equations with neural networks
**Submitting author:** @feiyu-chen96 (<a href="http://orcid.org/0000-0002-0476-9998">Feiyu Chen</a>)
**Repository:** <a href="https://github.com/odegym/neurodiffeq" target ="_blank">https://github.com/odegym/neurodiffeq</a>
**Version:** 1.0.0
