### 2019-11-01 #1859 @danielskatz
👋 @feiyu-chen96 - this paper may be a bit too long - see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
Please consider shortening it by moving some of the content to the software documentation.
