### 2019-11-01 #1859 @danielskatz
Also, I note that some of the citations to references are missing spaces before them.  And most of the references are missing DOIs
