### 2019-11-04 #1859 @drvinceknight
Yup, sounds like an interesting piece of software.
