### 2019-11-06 d12ad22
style/variable name changes
