### 2019-11-06 #1859 @feiyu-chen96
Hi @drvinceknight  Can I ask if there is a page limit? I'm thinking of keeping only the Summary section and move the rest to the documentation.
