### 2019-11-06 #1859 @drvinceknight
@feiyu-chen96 there are is no specific page limit but in general papers are short (around 2 or 3 pages long). There is some documentation here about what to include in the paper as well as links to accepted papers: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
