### 2019-11-07 9fd4d71
Removed case study section.
