+++ paper/paper.md
@@ -129,50 +129,4 @@ effort required during the training of the ANN[@mcfall2009artificial]. ``DeepXDE
 conditions. ``PyDEns`` uses a variation of the second approach to impose initial/boundary conditions. ``NeuroDiffEq``, the
 software described herein, employs the second approach. 
 
-## Case Studies
-
-Here we present two differential equations and how can they be solved by ``NeuroDiffEq``. In both cases, we are using 
-a fully connected neural network with one hidden layer of 32 nodes.
-
-### Lotka–Volterra equations
-
-The Lotka–Volterra equations are a system of first-order, nonlinear ODEs that have been used to model predator-prey dynamics
-in biological systems as well as problems in chemical kinetics.  They are given by:
-$$
-\begin{aligned}
-\frac{dx(t)}{dt} &= \alpha x(t) - \beta x(t)y(t), \quad x\left(0\right) = x_{0} \\
-\frac{dy(t)}{dt} &= \delta x(t)y(t) - \gamma y(t), \quad y\left(0\right) = y_{0}.
-\end{aligned}
-$$
-The time-evolution of the population of the prey and predator are given by $x$ and $y$, respectively, with $x_{0}$ and
-$y_{0}$ the initial populations. The coupling parameters $\alpha$, $\beta$, $\delta$ and $\gamma$ describe the interaction of
-the two species. Let $\alpha = \beta = \delta = \gamma = 1$, $x_{0} = 1.5$, and $y_{0} = 1.0$. For comparison purposes, we
-solve this problem numerically with ``scipy`` and ``NeuroDiffEq``.  Figure 1 compares the predator and prey populations from
-the numerical integrator and the neural network.  The solutions are qualitatively the same. 
-
-![case1](case1.png)
-*Figure 1: Comparing numerical and ANN-based solutions of Lotka–Volterra equations.* 
-
-### Poisson's equation
-
-Poisson's equation is a second-order linear PDE. It can be used to describe the potential field caused by a given charge 
-or mass density distribution. In two dimensional Cartesian coordinates, it takes the form:
-$$
-\left(\frac{\partial^2}{\partial x^2} + \frac{\partial^2}{\partial y^2}\right)u(x, y) = f(x, y).
-$$
-We solve the equation in the domain $(x, y) \in (0, 1) \times (0, 1)$ with homogeneous Dirichlet boundary conditions, 
-$$
-u(x, 0) = u(x, 1) = u(0, y) = u(1, y) = 0.
-$$
-With $f(x, y) = 2x(y-1)(y-2x+xy+2)e^{x-y}$ the analytical solution is
-$$
-u(x, y) = x(1-x)y(1-y)e^{x-y}.
-$$
-Figure 2 presents contours of the neural network solution (left), the analytical solution (middle), and the error between the
-analytical and neural network solution (right).  The largest error in the neural network solution is around $6\cdot 10^{-5}$.
-
-![case2](case2.png)
-*Figure 2: Comparing analytical and ANN-based solutions of Poisson's equation.* 
-
-
 # References
