+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={2-3},
   pages={667--690},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1063/1.4992752}
 }
 
 
@@ -17,7 +18,8 @@
   volume={2},
   pages={359--366},
   year={1989},
-  publisher={Pergamon Press}
+  publisher={Pergamon Press},
+  doi={10.1016/0893-6080(89)90020-8}
 }
 
 @article{lagaris1998artificial,
@@ -28,7 +30,8 @@
   number={5},
   pages={987--1000},
   year={1998},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/72.712178}
 }
 
 @article{mcfall2009artificial,
@@ -39,7 +42,8 @@
   number={8},
   pages={1221--1233},
   year={2009},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TNN.2009.2020735}
 }
 @article{lu2019deepxde,
   title={DeepXDE: A deep learning library for solving differential equations},
@@ -74,7 +78,8 @@
   booktitle={Proceedings of the Practice and Experience in Advanced Research Computing 2017 on Sustainability, Success and Impact},
   pages={70},
   year={2017},
-  organization={ACM}
+  organization={ACM},
+  doi={10.1145/3093338.3104151}
 }
 
 @misc{tensorflow2015-whitepaper,
