### 2019-11-07 4c647f6
Removed several sentences.
