+++ neurodiffeq/pde.py
@@ -10,6 +10,16 @@ from .networks import FCNN
 from .neurodiffeq import diff
 
 
+def _nn_output_2input(net, xs, ys):
+    xys = torch.cat((xs, ys), 1)
+    return net(xys)
+
+
+class NoCondition2D:
+    def enforce(self, net, x, y):
+        return _nn_output_2input(net, x, y)
+
+
 class DirichletBVP2D:
     """An Dirichlet boundary value problem on a 2-D orthogonal box where :math:`x\\in[x_0, x_1]` and :math:`y\\in[y_0, y_1]`
         We are solving :math:`u(x, t)` given:
@@ -59,8 +69,7 @@ class DirichletBVP2D:
             .. note::
                 `enforce` is meant to be called by the function `solve2D`.
         """
-        xys = torch.cat((x, y), 1)
-        u = net(xys)
+        u = _nn_output_2input(net, x, y)
         x_tilde = (x-self.x_min) / (self.x_max-self.x_min)
         y_tilde = (y-self.y_min) / (self.y_max-self.y_min)
         Axy = (1-x_tilde)*self.x_min_val(y) + x_tilde*self.x_max_val(y) + \
@@ -126,8 +135,7 @@ class IBVP1D:
             raise NotImplementedError('Sorry, this boundary condition is not implemented.')
 
     def _enforce_dd(self, net, x, t):
-        xts = torch.cat((x, t), 1)
-        uxt = net(xts)
+        uxt = _nn_output_2input(net, x, t)
 
         t_ones = torch.ones_like(t, requires_grad=True)
         t_ones_min = self.t_min * t_ones
@@ -141,15 +149,13 @@ class IBVP1D:
         return Axt + x_tilde * (1 - x_tilde) * (1 - torch.exp(-t_tilde)) * uxt
 
     def _enforce_dn(self, net, x, t):
-        xts = torch.cat((x, t), 1)
-        uxt = net(xts)
+        uxt = _nn_output_2input(net, x, t)
 
         x_ones = torch.ones_like(x, requires_grad=True)
         t_ones = torch.ones_like(t, requires_grad=True)
         x_ones_max = self.x_max * x_ones
         t_ones_min = self.t_min * t_ones
-        xmaxts  = torch.cat((x_ones_max, t), 1)
-        uxmaxt  = net(xmaxts)
+        uxmaxt = _nn_output_2input(net, x_ones_max, t)
 
         x_tilde = (x-self.x_min) / (self.x_max-self.x_min)
         t_tilde = t-self.t_min
@@ -161,15 +167,13 @@ class IBVP1D:
         )
 
     def _enforce_nd(self, net, x, t):
-        xts = torch.cat((x, t), 1)
-        uxt = net(xts)
+        uxt = _nn_output_2input(net, x, t)
 
         x_ones = torch.ones_like(x, requires_grad=True)
         t_ones = torch.ones_like(t, requires_grad=True)
         x_ones_min = self.x_min * x_ones
         t_ones_min = self.t_min * t_ones
-        xmints = torch.cat((x_ones_min, t), 1)
-        uxmint = net(xmints)
+        uxmint = _nn_output_2input(net, x_ones_min, t)
 
         x_tilde = (x - self.x_min) / (self.x_max - self.x_min)
         t_tilde = t - self.t_min
@@ -181,18 +185,15 @@ class IBVP1D:
         )
 
     def _enforce_nn(self, net, x, t):
-        xts = torch.cat((x, t), 1)
-        uxt = net(xts)
+        uxt = _nn_output_2input(net, x, t)
 
         x_ones = torch.ones_like(x, requires_grad=True)
[…]
