### 2019-11-08 b335d78
Fixed space before citation.
