### 2019-11-08 #1859 @feiyu-chen96
@drvinceknight I just shorten the paper and added DOI and space for citations. Could you check the updated version?
