### 2019-11-08 5250524
Created separate class for Solution.
