+++ neurodiffeq/ode.py
@@ -6,6 +6,7 @@ import torch.nn as nn
 import torch.optim as optim
 
 from .networks import FCNN
+from copy import deepcopy
 
 
 class IVP:
@@ -158,7 +159,7 @@ class Monitor:
         :param nets: The neural networks that approximates the ODE (system).
         :type nets: list[`torch.nn.Module`]
         :param conditions: The initial/boundary conditions of the ODE (system).
-        :type conditions: list [`neurodiff.ode.IVP` or `neurodiff.ode.DirichletBVP`]
+        :type conditions: list[`neurodiffeq.ode.IVP` or `neurodiffeq.ode.DirichletBVP`]
         :param loss_history: The history of training loss and validation loss. The 'train' entry is a list of training loss and 'valid' entry is a list of validation loss.
         :type loss_history: dict['train': list[float], 'valid': list[float]]
 
@@ -205,7 +206,7 @@ def solve(
         then `ode` should be a function that maps :math:`(x, t)` to :math:`F(x, t)`.
     :type ode: function
     :param condition: The initial/boundary condition.
-    :type condition: `neurodiff.ode.IVP` or `neurodiff.ode.DirichletBVP`
+    :type condition: `neurodiffeq.ode.IVP` or `neurodiffeq.ode.DirichletBVP`
     :param net: The neural network used to approximate the solution, defaults to None.
     :type net: `torch.nn.Module`, optional
     :param t_min: The lower bound of the domain (t) on which the ODE is solved.
@@ -213,11 +214,11 @@ def solve(
     :param t_max: The upper bound of the domain (t) on which the ODE is solved.
     :type t_max: float
     :param train_generator: The example generator to generate 1-D training points, default to None.
-    :type train_generator: `neurodiff.ode.ExampleGenerator`, optional
+    :type train_generator: `neurodiffeq.ode.ExampleGenerator`, optional
     :param shuffle: Whether to shuffle the training examples every epoch, defaults to True.
     :type shuffle: bool, optional
     :param valid_generator: The example generator to generate 1-D validation points, default to None.
-    :type valid_generator: `neurodiff.ode.ExampleGenerator`, optional
+    :type valid_generator: `neurodiffeq.ode.ExampleGenerator`, optional
     :param optimizer: The optimization method to use for training, defaults to None.
     :type optimizer: `torch.optim.Optimizer`, optional
     :param criterion: The loss function to use for training, defaults to None.
@@ -234,10 +235,10 @@ def solve(
     :type return_best: bool, optional
     :return: The solution of the ODE. The history of training loss and validation loss.
         Optionally, the nets, conditions, training generator, validation generator, optimizer and loss function.
-    :rtype: tuple[function, dict]; or tuple[function, dict, dict]
+    :rtype: tuple[`neurodiffeq.ode.Solution`, dict]; or tuple[`neurodiffeq.ode.Solution`, dict, dict]
     """
     nets = None if not net else [net]
-    returned_tuple = solve_system(
+    return solve_system(
         ode_system=lambda x, t: [ode(x, t)], conditions=[condition],
         t_min=t_min, t_max=t_max, nets=nets,
         train_generator=train_generator, shuffle=shuffle, valid_generator=valid_generator,
@@ -245,15 +246,6 @@ def solve(
         max_epochs=max_epochs, monitor=monitor, return_internal=return_internal,
         return_best=return_best
     )
-
-    def solution_wrapped(ts, as_type='tf'):
-        return solution(ts, as_type)[0]
-    if return_internal:
-        solution, loss_history, internal = returned_tuple
-        return solution_wrapped, loss_history, internal
-    else:
-        solution, loss_history = returned_tuple
-        return solution_wrapped, loss_history
 
 
[…]
