### 2019-11-08 621e564
Style changes + Added NoCondition
