### 2019-11-08 b07d701
Added Solution class and return_best option for PDEs.
