+++ neurodiffeq/pde.py
@@ -8,6 +8,7 @@ import matplotlib.animation as animation
 
 from .networks import FCNN
 from .neurodiffeq import diff
+from copy import deepcopy
 
 
 def _nn_output_2input(net, xs, ys):
@@ -16,7 +17,8 @@ def _nn_output_2input(net, xs, ys):
 
 
 class NoCondition2D:
-    def enforce(self, net, x, y):
+    @staticmethod
+    def enforce(net, x, y):
         return _nn_output_2input(net, x, y)
 
 
@@ -340,7 +342,7 @@ def solve2D(
         pde, condition, xy_min, xy_max,
         net=None, train_generator=None, shuffle=True, valid_generator=None, optimizer=None, criterion=None, batch_size=16,
         max_epochs=1000,
-        monitor=None, return_internal=False
+        monitor=None, return_internal=False, return_best=False
 ):
     """Train a neural network to solve a PDE with 2 independent variables.
 
@@ -373,30 +375,21 @@ def solve2D(
     :type monitor: `neurodiffeq.pde.Monitor2D`, optional
     :param return_internal: Whether to return the nets, conditions, training generator, validation generator, optimizer and loss function, defaults to False.
     :type return_internal: bool, optional
+    :param return_best: Whether to return the nets that achieved the lowest validation loss, defaults to False.
+    :type return_best: bool, optional
     :return: The solution of the PDE. The history of training loss and validation loss.
         Optionally, the nets, conditions, training generator, validation generator, optimizer and loss function.
         The solution is a function that has the signature `solution(xs, ys, as_type)`.
-        `xs (torch.tensor)` and `ys (torch.tensor)` are the points on which :math:`u(x, y)` is evaluated.
-        `as_type (str)` indicates whether the returned value is a `torch.tensor` ('tf') or `numpy.array` ('np').
-    :rtype: tuple[function, dict]; or tuple[function, dict, dict]
+    :rtype: tuple[`neurodiffeq.pde.Solution`, dict]; or tuple[`neurodiffeq.pde.Solution`, dict, dict]
     """
     nets = None if not net else [net]
-    returned_tuple = solve2D_system(
+    return solve2D_system(
         pde_system=lambda u, x, y: [pde(u, x, y)], conditions=[condition],
         xy_min=xy_min, xy_max=xy_max, nets=nets,
         train_generator=train_generator, shuffle=shuffle, valid_generator=valid_generator,
         optimizer=optimizer, criterion=criterion, batch_size=batch_size,
-        max_epochs=max_epochs, monitor=monitor, return_internal=return_internal
+        max_epochs=max_epochs, monitor=monitor, return_internal=return_internal, return_best=return_best
     )
-
-    def solution_wrapped(xs, ys, as_type='tf'):
-        return solution(xs, ys, as_type)[0]
-    if return_internal:
-        solution, loss_history, internal = returned_tuple
-        return solution_wrapped, loss_history, internal
-    else:
-        solution, loss_history = returned_tuple
-        return solution_wrapped, loss_history
 
 
 def solve2D_system(
@@ -404,7 +397,7 @@ def solve2D_system(
         nets=None, train_generator=None, shuffle=True, valid_generator=None,
         optimizer=None, criterion=None, batch_size=16,
         max_epochs=1000,
-        monitor=None, return_internal=False
+        monitor=None, return_internal=False, return_best=False
 ):
     # default values
     n_dependent_vars = len(conditions)
@@ -440,6 +433,8 @@ def solve2D_system(
     valid_zeros = torch.zeros(n_examples_valid)
 
     loss_history = {'train': [], 'valid': []}
+    valid_loss_epoch_min = np.inf
+    solution_min = None
 
     for epoch in range(max_epochs):
[…]
