### 2019-11-08 baa49d7
Updated doc strings
