### 2019-11-10 d3b2a75
Made the input shape and output shape the same in Solutions.
