+++ neurodiffeq/ode.py
@@ -436,7 +436,7 @@ class Solution:
             `torch.tensor` or `numpy.array` (when there is only one dependent variable)
         """
         if not isinstance(ts, torch.Tensor):
-            ts = torch.tensor([ts], dtype=torch.float32)
+            ts = torch.tensor(ts, dtype=torch.float32)
         ts = ts.reshape(-1, 1)
         if as_type not in ('tf', 'np'):
             raise ValueError("The valid return types are 'tf' and 'np'.")
