### 2019-11-12 #1859 @drvinceknight
Thanks @feiyu-chen96, I think this looks more like a JOSS paper now. I will start looking to assign reviewers. On that note do you have any you might suggest?
