### 2019-11-16 #1859 @drvinceknight
@arghdos, @ysimillides, @Jgoldfar would you be able to able to review this paper for JOSS? There is information about what that involves here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and I'm happy to answer any questions you may have.
