### 2019-11-18 #1859 @jgoldfar
I will have time to review more effectively next January; please keep me in mind at that time!

> On Nov 16, 2019, at 3:04 PM, Vince Knight <notifications@github.com> wrote:
>
> ﻿
> @arghdos, @ysimillides, @jgoldfar would you be able to able to review this paper for JOSS? There is information about what that involves here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and I'm happy to answer any questions you may have.
>
