### 2019-11-28 #1859 @drvinceknight
@lheagy @dawbarton mancellin would you be able to review this paper for JOSS? Your specified areas of interest include differential equations which this is a tool for so hopefully you'll find it interesting. It implements solving differential equations using neural networks.

If you haven't reviewed for JOSS before, there is information about what that involves here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and I'm super happy to answer any questions you may have.
