### 2019-12-02 #1859 @drvinceknight
@marcharper this is the paper I mentioned on email, would you be interested in reviewing it?

If you haven't reviewed for JOSS before, there is information about what that involves here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and I'm super happy to answer any questions you may have.
