+++ experiments/profiled.py
@@ -1,0 +1,206 @@
+import numpy as np
+import torch
+import matplotlib.pyplot as plt
+from neurodiffeq import diff
+from neurodiffeq.networks import FCNN
+from neurodiffeq.ode import IVP, Monitor, solve_system
+from neurodiffeq.pde import solve2D, ExampleGenerator2D
+
+class Point:
+    
+    def __repr__(self):
+        return f'Point({self.loc})'
+    
+    def __init__(self, loc):
+        self.loc = tuple(float(d) for d in loc)
+        self.dim = len(loc)
+
+class DirichletControlPoint(Point):
+    
+    def __repr__(self):
+        return f'DirichletControlPoint({self.loc}, val={self.val})'
+    
+    def __init__(self, loc, val):
+        super().__init__(loc)
+        self.val = val
+
+class Condition:
+    
+    def enforce(self, net, *dimensions):
+        raise NotImplemented
+    
+    @staticmethod
+    def _nn_output(net, *dimensions):
+        original_shape = dimensions[0].shape
+        output = net(torch.cat(dimensions, 1))
+        return output.reshape(original_shape)
+
+class CustomDirichletBoundaryCondition(Condition):
+    
+    def __init__(self, dirichlet_control_points, center_point):
+        # drop deplicates and sort 'clockwise' (while not changing the original copy)
+        dirichlet_control_points = self._clean_control_points(dirichlet_control_points, center_point)
+        # fit Dirichlet dummy solution (A_D(x) in MacFall's paper)
+        self.a_d = _Interpolator.fit_dirichlet_dummy_solution(dirichlet_control_points)
+        # fit Dirichlet length factor (L_D(x) in MacFall's paper)
+        self.l_d = _Interpolator.fit_dirichlet_length_factor(dirichlet_control_points)
+    
+    def enforce(self, net, *dimensions):
+        # enforce Dirichlet boundary condition u_t(x) = A_D(x) + L_D(x)u_N(x)
+        return self.a_d(*dimensions) + self.l_d(*dimensions)*self._nn_output(net, *dimensions)
+    
+    @staticmethod
+    def _clean_control_points(control_points, center_point):
+        # remove the control points that are defined more than once
+        locs = set()
+        unique_control_points = []
+        for cp in control_points:
+            if cp.loc not in locs:
+                locs.add(cp.loc)
+                unique_control_points.append(cp)
+
+        # sort the control points 'clockwise' (from 0 to -2pi)
+        def clockwise(cp):
+            return (
+                1 if cp.loc[1]-center_point.loc[1] > 0 else -1,
+                (cp.loc[0]-center_point.loc[0])/(cp.loc[1]-center_point.loc[1]-1e-9)
+            )
+        unique_control_points.sort(key=clockwise)
+        return unique_control_points
+
+class _Interpolator:
+
+    @staticmethod
+    def fit_dirichlet_dummy_solution(dirichlet_control_points):
+        # specify input and output of thin plate spline
+        from_points = dirichlet_control_points
+        to_values = [dcp.val for dcp in dirichlet_control_points]
+        # fit thin plate spline and save coefficients
+        coefs = _Interpolator.solve_thin_plate_spline(from_points, to_values)
+        # define closure that map arbitrary points x (represented by torch.tensor 
+        # with shape [#point, #dimension]) to A_D(x)
+        def a_d(*dimensions):
+            return _Interpolator.interpolate_by_thin_plate_spline(coefs, from_points, *dimensions)
+        return a_d
+        
+    @staticmethod
+    def fit_dirichlet_length_factor(dirichlet_control_points, radius=0.5):
+        # specify input and output of thin plate spline
[…]
