### 2019-12-14 837dd50
Implemented functionality for steady-state PDEs in spherical coordinates
