+++ neurodiffeq/pde_spherical.py
@@ -1,0 +1,527 @@
+import torch
+import torch.optim as optim
+import torch.nn as nn
+
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib.animation as animation
+
+from .networks import FCNN
+from .neurodiffeq import diff
+from copy import deepcopy
+
+
+def _nn_output_spherical_input(net, rs, thetas, phis):
+    points = torch.cat((rs, thetas, phis), 1)
+    return net(points)
+
+
+class NoConditionSpherical:
+    @staticmethod
+    def enforce(net, r, theta, phi):
+        return _nn_output_spherical_input(net, r, theta, phi)
+
+
+class ExampleGenerator3D:
+    """An example generator for generating 3-D training points. NOT TO BE CONFUSED with `ExampleGeneratorSpherical`
+        :param grid: The discretization of the 3 dimensions, if we want to generate points on a :math:`m \\times n \\times k` grid, then `grid` is `(m, n, k)`, defaults to `(10, 10, 10)`.
+        :type grid: tuple[int, int, int], optional
+        :param xyz_min: The lower bound of 3 dimensions, if we only care about :math:`x \\geq x_0`, :math:`y \\geq y_0`, and :math:`z \\geq z_0` then `xyz_min` is `(x_0, y_0, z_0)`, defaults to `(0.0, 0.0, 0.0)`.
+        :type xyz_min: tuple[float, float, float], optional
+        :param xyz_max: The upper bound of 3 dimensions, if we only care about :math:`x \\leq x_1`, :math:`y \\leq y_1`, and :math:`z \\leq z_1` then `xyz_max` is `(x_1, y_1, z_1)`, defaults to `(1.0, 1.0, 1.0)`.
+        :type xyz_max: tuple[float, float, float], optional
+        :param method: The distribution of the 2-D points generated.
+            If set to 'equally-spaced', the points will be fixed to the grid specified.
+            If set to 'equally-spaced-noisy', a normal noise will be added to the previously mentioned set of points, defaults to 'equally-spaced-noisy'.
+        :type method: str, optional
+        :raises ValueError: When provided with an unknown method.
+    """
+
+    def __init__(self, grid=(10, 10, 10), xyz_min=(0.0, 0.0, 0.0), xyz_max=(1.0, 1.0, 1.0),
+                 method='equally-spaced-noisy'):
+        r"""Initializer method
+
+        .. note::
+            A instance method `get_examples` is dynamically created to generate 2-D training points. It will be called by the function `solve2D`.
+        """
+        self.size = grid[0] * grid[1] * grid[2]
+
+        x = torch.linspace(xyz_min[0], xyz_max[0], grid[0], requires_grad=True)
+        y = torch.linspace(xyz_min[1], xyz_max[1], grid[1], requires_grad=True)
+        z = torch.linspace(xyz_min[2], xyz_max[2], grid[2], requires_grad=True)
+        grid_x, grid_y, grid_z = torch.meshgrid(x, y, z)
+        self.grid_x, self.grid_y, self.grid_z = grid_x.flatten(), grid_y.flatten(), grid_z.flatten()
+
+        if method == 'equally-spaced':
+            self.get_examples = lambda: (self.grid_x, self.grid_y, self.grid_z)
+        elif method == 'equally-spaced-noisy':
+            self.noise_xmean = torch.zeros(self.size)
+            self.noise_ymean = torch.zeros(self.size)
+            self.noise_zmean = torch.zeros(self.size)
+            self.noise_xstd = torch.ones(self.size) * ((xyz_max[0] - xyz_min[0]) / grid[0]) / 4.0
[…]
