### 2019-12-15 6745e18
implemented and tested infinity BC, added a base class for spherical BCs
