+++ neurodiffeq/pde_spherical.py
@@ -16,9 +16,16 @@ def _nn_output_spherical_input(net, rs, thetas, phis):
     return net(points)
 
 
-class NoConditionSpherical:
-    @staticmethod
-    def enforce(net, r, theta, phi):
+class BaseBVPSpherical:
+    def enforce(self, net, r, theta, phi):
+        raise NotImplementedError(f"Abstract {self.__class__.__name__} cannot be enforced")
+
+
+class NoConditionSpherical(BaseBVPSpherical):
+    def __init__(self):
+        pass
+
+    def enforce(self, net, r, theta, phi):
         return _nn_output_spherical_input(net, r, theta, phi)
 
 
@@ -130,7 +137,7 @@ class ExampleGeneratorSpherical:
         return r, theta, phi
 
 
-class DirichletBVPSpherical:
+class DirichletBVPSpherical(BaseBVPSpherical):
     """Dirichlet boundary condition for the interior and exterior boundary of the sphere, where the interior boundary is not necessarily a point
         We are solving :math:`u(t)` given :math:`u(r, \\theta, \\phi)\\bigg|_{r = r_0} = f(\\theta, \\phi)` and :math:`u(r, \\theta, \\phi)\\bigg|_{r = r_1} = g(\\theta, \\phi)`
 
@@ -166,7 +173,7 @@ class DirichletBVPSpherical:
 
 
         .. note::
-            `enforce` is meant to be called by the function `solve` and `solve_system`.
+            `enforce` is meant to be called by the function `solve_spherical` and `solve_spherical_system`.
         """
         u = _nn_output_spherical_input(net, r, theta, phi)
         r_tilde = (r - self.r_0) / (self.r_1 - self.r_0)
@@ -176,13 +183,58 @@ class DirichletBVPSpherical:
                (1. - torch.exp((1 - r_tilde) * r_tilde)) * u
 
 
+class InfDirichletBVPSpherical(BaseBVPSpherical):
+    """Similar to `DirichletBVPSpherical`; only difference is we are considering :math:`g(\\theta, \\phi)` as :math:`r_1 \\to \\infty`, so `r_1` doesn't need to be specified
+        We are solving :math:`u(t)` given :math:`u(r, \\theta, \\phi)\\bigg|_{r = r_0} = f(\\theta, \\phi)` and :math:`\\lim_{r \\to \\infty} u(r, \\theta, \\phi) = g(\\theta, \\phi)`
+
+    :param r_0: The radius of the interior boundary. When r_0 = 0, the interior boundary is collapsed to a single point (center of the ball)
+    :type r_0: float
+    :param f: The value of :math:u on the interior boundary. :math:`u(r, \\theta, \\phi)\\bigg|_{r = r_0} = f(\\theta, \\phi)`.
+    :type f: function
+    :param g: The value of :math:u on the exterior boundary. :math:`u(r, \\theta, \\phi)\\bigg|_{r = r_1} = g(\\theta, \\phi)`.
+    :type g: function
+    :param order: The smallest :math:k that guarantees :math:`\\lim_{r \\to +\\infty} u(r, \\theta, \\phi) e^{-k r} = 0`, defaults to 1
+    :type order: int or float, optional
+    """
+
+    def __init__(self, r_0, f, g, order=1):
+        self.r_0 = r_0
+        self.f = f
+        self.g = g
+        self.order = order
+
+    def enforce(self, net, r, theta, phi):
+        r"""Enforce the output of a neural network to satisfy the boundary condition.
+
+        :param net: The neural network that approximates the ODE.
+        :type net: `torch.nn.Module`
+        :param r: The radii of points where the neural network output is evaluated.
+        :type r: `torch.tensor`
[…]
