### 2019-12-15 a6fbbfe
added test case for Poisson Eq; visualized MSE against analytic solution
