+++ neurodiffeq/pde_spherical.py
@@ -283,7 +283,7 @@ class SolutionSpherical:
 
 def solve_spherical(
         pde, condition, r_min, r_max,
-        net=None, train_generator=None, shuffle=True, valid_generator=None,
+        net=None, train_generator=None, shuffle=True, valid_generator=None, analytic_solution=None,
         optimizer=None, criterion=None, batch_size=16, max_epochs=1000,
         monitor=None, return_internal=False, return_best=False
 ):
@@ -306,6 +306,8 @@ def solve_spherical(
         :type shuffle: bool, optional
         :param valid_generator: The example generator to generate 3-D validation points, default to None.
         :type valid_generator: `neurodiffeq.pde_spherical.ExampleGeneratorSpherical`, optional
+        :param analytic_solution: analytic solution to the pde system, used for testing purposes; should map (rs, thetas, phis) to u
+        :type analytic_solution: function
         :param optimizer: The optimization method to use for training, defaults to None.
         :type optimizer: `torch.optim.Optimizer`, optional
         :param criterion: The loss function to use for training, defaults to None.
@@ -321,29 +323,32 @@ def solve_spherical(
         :param return_best: Whether to return the nets that achieved the lowest validation loss, defaults to False.
         :type return_best: bool, optional
         :return: The solution of the PDE. The history of training loss and validation loss.
-            Optionally, the nets, conditions, training generator, validation generator, optimizer and loss function.
+            Optionally, MSE against analytic solution, the nets, conditions, training generator, validation generator, optimizer and loss function.
             The solution is a function that has the signature `solution(xs, ys, as_type)`.
-        :rtype: tuple[`neurodiffeq.pde_spherical.SolutionSpherical`, dict]; or tuple[`neurodiffeq.pde_spherical.SolutionSpherical`, dict, dict]
+        :rtype: tuple[`neurodiffeq.pde_spherical.SolutionSpherical`, dict]; or tuple[`neurodiffeq.pde_spherical.SolutionSpherical`, dict, dict]; or tuple[`neurodiffeq.pde_spherical.SolutionSpherical`, dict, dict, dict]
         """
 
     pde_sytem = lambda u, r, theta, phi: [pde(u, r, theta, phi)]
     conditions = [condition]
     nets = [net] if net is not None else None
+    if analytic_solution is None:
+        analytic_solutions = None
+    else:
+        analytic_solutions = lambda r, theta, phi: [analytic_solution(r, theta, phi)]
 
     return solve_spherical_system(
         pde_system=pde_sytem, conditions=conditions, r_min=r_min, r_max=r_max,
         nets=nets, train_generator=train_generator, shuffle=shuffle, valid_generator=valid_generator,
-        optimizer=optimizer, criterion=criterion, batch_size=batch_size, max_epochs=max_epochs,
-        monitor=monitor, return_internal=return_internal, return_best=return_best,
+        analytic_solutions=analytic_solutions, optimizer=optimizer, criterion=criterion, batch_size=batch_size,
+        max_epochs=max_epochs, monitor=monitor, return_internal=return_internal, return_best=return_best,
     )
 
 
 def solve_spherical_system(
         pde_system, conditions, r_min, r_max,
-        nets=None, train_generator=None, shuffle=True, valid_generator=None,
+        nets=None, train_generator=None, shuffle=True, valid_generator=None, analytic_solutions=None,
         optimizer=None, criterion=None, batch_size=16,
-        max_epochs=1000,
-        monitor=None, return_internal=False, return_best=False
+        max_epochs=1000, monitor=None, return_internal=False, return_best=False
 ):
[…]
