### 2019-12-15 173d752
Updated profiling and timing experiments
