### 2019-12-15 8d01836
Applied command line plotting fix.
