### 2019-12-16 be8e4c1
increased tolerance margin and max_epochs for Poisson test
