### 2019-12-17 d9be7dc
modified the range of phi to [0, 2 pi) instead of (-pi, pi]
