### 2019-12-19 774aa27
support specifying only 1 end of the boundary instead of both
