### 2019-12-19 fd0df5b
add test case for potential gaussian charged density
refer to this link for more info: https://en.wikipedia.org/wiki/Poisson%27s_equation
