### 2019-12-20 16532e1
added viz of solution (colorbar not working right)
