+++ neurodiffeq/pde_spherical.py
@@ -3,6 +3,8 @@ import torch.optim as optim
 import torch.nn as nn
 
 import numpy as np
+import pandas as pd
+import seaborn as sns
 import matplotlib.pyplot as plt
 import matplotlib.animation as animation
 
@@ -545,15 +547,30 @@ class MonitorSpherical:
         self.axs = []  # subplots
         self.cbs = []  # color bars
         # input for neural network
-        gen = ExampleGeneratorSpherical(256, r_min=r_min, r_max=r_max, method='equally-spaced-noisy')
-        rs_ann, thetas_ann, phis_ann = gen.get_examples()
-        self.rs_ann = rs_ann.reshape(-1, 1)
-        self.thetas_ann = thetas_ann.reshape(-1, 1)
-        self.phis_ann = phis_ann.reshape(-1, 1)
-        # self.xy_ann = torch.cat((self.xs_ann, self.ys_ann), 1)
+        gen = ExampleGenerator3D(
+            grid=(10, 10, 10),
+            xyz_min=(r_min, 0., 0.),
+            xyz_max=(r_max, np.pi, 2 * np.pi),
+            method='equally-spaced'
+        )
+        rs, thetas, phis = gen.get_examples()
+
+        th = thetas.reshape(10, 10, 10)[0, :, 0].detach().numpy()
+        ph = phis.reshape(10, 10, 10)[0, 0, :].detach().numpy()
+
+        self.rs = rs.reshape(-1, 1)
+        self.thetas = thetas.reshape(-1, 1)
+        self.phis = phis.reshape(-1, 1)
+        self.th = th
+        self.ph = ph
 
     def check(self, nets, conditions, loss_history, analytic_mse_history=None):
-        r"""Draw 2 plots: One shows the shape of the current solution (with heat map). The other shows the history training loss and validation loss.
+        r"""Draw (3n + 2) plots:
+             1) For each function u(r, phi, theta), there are 3 axes:
+                a) one ax for u-r curves grouped by phi
+                b) one ax for u-r curves grouped by theta
+                c) one ax for u-theta-phi contour heat map
+             2) Additionally, one ax for MSE against analytic solution, another for training and validation loss
 
         :param nets: The neural networks that approximates the PDE.
         :type nets: list [`torch.nn.Module`]
@@ -568,37 +585,79 @@ class MonitorSpherical:
             `check` is meant to be called by the function `solve2D`.
         """
 
+        # initialize the figure and axes here so that the Monitor knows the number of dependent variables and
+        # shape of the figure, number of the subplots, etc.
+        # Draw (3n + 2) plots:
+        #     1) For each function u(r, phi, theta), there are 3 axes:
+        #         a) one ax for u-r curves grouped by phi
+        #         b) one ax for u-r curves grouped by theta
+        #         c) one ax for u-theta-phi contour heat map
+        #     2) Additionally, one ax for MSE against analytic solution, another for training and validation loss
+        n_axs = len(nets) * 3 + 2
+        n_row = len(nets) + 1
+        n_col = 3
         if not self.fig:
-            # initialize the figure and axes here so that the Monitor knows the number of dependent variables and
-            # shape of the figure, number of the subplots, etc.
-            n_axs = 2  # one for MSE against analytic solution, the other for training and validation loss
-            n_row, n_col = (n_axs + 1) // 2, 2
-            self.fig = plt.figure(figsize=(20, 8 * n_row))
+            self.fig = plt.figure(figsize=(20, 6 * n_row))
             for i in range(n_axs):
                 self.axs.append(self.fig.add_subplot(n_row, n_col, i + 1))
             for i in range(len(nets)):
[…]
