### 2019-12-20 dcaffb0
switched to heatmap and turned on colorbar There might be a bug with colorbar removal for contourf when there are multiple subplots
