### 2019-12-21 f90f682
allowed specifying of variable names in MonitorSpherical
