### 2019-12-23 c499500
added method to create a new figure, useful for plotting best solution
