### 2019-12-26 26c60e4
added support for gpu version changed all .detach().numpy() to .detach().cpu().numpy()
