### 2020-01-07 532aba8
fix a bug when batch_size is not always the same
