### 2020-01-09 dd7f9f4
allow the user to specify the scale of the noise when generating training points + some clean up of code in ExampleGenerator, I dont understand why I wrote them in that funny way...
