### 2020-01-09 1e7d977
fix random seed
