### 2020-01-11 4c1c486
restructure code so that solution can be picked
