+++ experiments/custombc.py
@@ -39,13 +39,19 @@ class CustomDirichletBoundaryCondition(Condition):
         # drop deplicates and sort 'clockwise' (while not changing the original copy)
         dirichlet_control_points = self._clean_control_points(dirichlet_control_points, center_point)
         # fit Dirichlet dummy solution (A_D(x) in MacFall's paper)
-        self.a_d = _Interpolator.fit_dirichlet_dummy_solution(dirichlet_control_points)
+        self.a_d_interp = InterpolatorCreator.fit_dirichlet_dummy_solution(dirichlet_control_points)
         # fit Dirichlet length factor (L_D(x) in MacFall's paper)
-        self.l_d = _Interpolator.fit_dirichlet_length_factor(dirichlet_control_points)
+        self.l_d_interp = InterpolatorCreator.fit_dirichlet_length_factor(dirichlet_control_points)
+    
+    def a_d(self, *dimensions):
+        return self.a_d_interp.interpolate(*dimensions)
+    
+    def l_d(self, *dimensions):
+        return self.l_d_interp.interpolate(*dimensions)
     
     def enforce(self, net, *dimensions):
         # enforce Dirichlet boundary condition u_t(x) = A_D(x) + L_D(x)u_N(x)
-        return self.a_d(*dimensions) + self.l_d(*dimensions)*self._nn_output(net, *dimensions)
+        return self.a_d(*dimensions) + self.l_d(*dimensions) * self._nn_output(net, *dimensions)
     
     @staticmethod
     def _clean_control_points(control_points, center_point):
@@ -67,7 +73,7 @@ class CustomDirichletBoundaryCondition(Condition):
         return unique_control_points
 
 
-class _Interpolator:
+class InterpolatorCreator:
 
     @staticmethod
     def fit_dirichlet_dummy_solution(dirichlet_control_points):
@@ -75,37 +81,25 @@ class _Interpolator:
         from_points = dirichlet_control_points
         to_values = [dcp.val for dcp in dirichlet_control_points]
         # fit thin plate spline and save coefficients
-        coefs = _Interpolator.solve_thin_plate_spline(from_points, to_values)
-        # define closure that map arbitrary points x (represented by torch.tensor 
-        # with shape [#point, #dimension]) to A_D(x)
-        def a_d(*dimensions):
-            return _Interpolator.interpolate_by_thin_plate_spline(coefs, from_points, *dimensions)
-        return a_d
+        coefs = InterpolatorCreator._solve_thin_plate_spline(from_points, to_values)
+        return ADInterpolator(coefs, dirichlet_control_points)
         
     @staticmethod
     def fit_dirichlet_length_factor(dirichlet_control_points, radius=0.5):
         # specify input and output of thin plate spline
         from_points = dirichlet_control_points
-        to_points = _Interpolator._create_circular_targets(dirichlet_control_points, radius)
+        to_points = InterpolatorCreator._create_circular_targets(dirichlet_control_points, radius)
         n_dim = to_points[0].dim
         to_values_each_dim = [[tp.loc[i] for tp in to_points] for i in range(n_dim)]
         # fit thin plate spline and save coefficients
         coefs_each_dim = [
-            _Interpolator.solve_thin_plate_spline(from_points, to_values)
+            InterpolatorCreator._solve_thin_plate_spline(from_points, to_values)
             for to_values in to_values_each_dim
         ]
-        # define closure that map arbitrary points x (represented by torch.tensor 
-        # with shape [#point, #dimension]) to L_D(x)
-        def l_d(*dimensions):
-            dimensions_mapped = [
[…]
