### 2020-01-12 #1931 @marcharper
I am declaring a potential conflict (IMO very minor)  -- this library uses torch/pytorch, my current employer creates a competing package (tensorflow) which I have not contributed to.

Authors: Feiyu Chen appears to have written the vast majority of the code along with another contributor, David Sondak. The two other authors do not appear to have contributed to the code AFAICT and are collaborators with the other authors. A very recent contributor (Johnny-Wish) is not listed as an author.

Installation: OK, but
- the given installation command does not install the requirements
- I additionally needed to run `pip install torch` to be able to import the package in a python interpreter
- I further needed to install jupyter, matplotlib, scipy in the venv to run examples

Examples: The examples run fine on a decent laptop.

Running tests: Tests are present but there are no instructions on how to run tests (at least not in a clearly visible place)
- Additional packages needed for running tests: `pip install pandas pytest seaborn`
- Tests can then be run with `/.../env/bin/pytest neurodiffpy`
- The tests open a lot of images and redraw them repeatedly; recommend closing them before moving on to subsequent tests / only displaying once

Contribution documentation is present but minimal. As noted there is a new recent contributor so seemingly the documentation is sufficient.

Citations and references look fine. It may be worth citing some of the more recent work on using neural networks to solve differential equations, if they are relevant. I defer to the authors choices.
