### 2020-01-14 d84dba6
fix a bug in sorting clockwise
