### 2020-01-16 25215f7
not show plots during testing
