### 2020-01-16 e047264
not call plt.pause when using non-GUI backend, so that no warning during testing
