+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 sphinx
 mock
+ipython
 nbsphinx
 sphinx_rtd_theme
 recommonmark
