### 2020-01-20 ed8e7e6
results of autograd profiler, looks normal
