### 2020-01-22 b1e4508
extract subroutines from solve2D_system, the function was too longand hard to modify
