### 2020-01-22 7b8efed
in solve2D, make xy_min and xy_max optional when train_generator and test_generator is specified
