+++ tests/test_pde.py
@@ -1,5 +1,4 @@
 import numpy as np
-import torch
 from numpy import isclose
 import matplotlib
 matplotlib.use('Agg') # use a non-GUI backend, so plots are not shown during testing
@@ -28,11 +27,11 @@ def test_monitor():
 
     net = FCNN(n_input_units=2, n_hidden_units=32, n_hidden_layers=1)
     solution_neural_net_laplace, _ = solve2D(
-        pde=laplace, condition=bc, xy_min=[0, 0], xy_max=[1, 1],
+        pde=laplace, condition=bc, xy_min=(0, 0), xy_max=(1, 1),
         net=net, max_epochs=3,
-        train_generator=ExampleGenerator2D([32, 32], [0, 0], [1, 1], method='equally-spaced-noisy'),
+        train_generator=ExampleGenerator2D((32, 32), (0, 0), (1, 1), method='equally-spaced-noisy'),
         batch_size=64,
-        monitor=Monitor2D(check_every=1, xy_min=[0, 0], xy_max=[1, 1])
+        monitor=Monitor2D(check_every=1, xy_min=(0, 0), xy_max=(1, 1))
     )
     print('Monitor test passed.')
 
@@ -48,27 +47,40 @@ def test_train_generator():
 
     net = FCNN(n_input_units=2, n_hidden_units=32, n_hidden_layers=1)
     solution_neural_net_laplace, _ = solve2D(
-        pde=laplace, condition=bc, xy_min=[0, 0], xy_max=[1, 1],
+        pde=laplace, condition=bc, xy_min=(0, 0), xy_max=(1, 1),
         net=net, max_epochs=3,
-        train_generator=ExampleGenerator2D([32, 32], [0, 0], [1, 1], method='equally-spaced-noisy'),
+        train_generator=ExampleGenerator2D((32, 32), (0, 0), (1, 1), method='equally-spaced-noisy'),
         batch_size=64,
-        monitor=Monitor2D(check_every=1, xy_min=[0, 0], xy_max=[1, 1])
+        monitor=Monitor2D(check_every=1, xy_min=(0, 0), xy_max=(1, 1))
     )
 
-    train_gen = ExampleGenerator2D([32, 32], [0, 0], [1, 1], method='equally-spaced')
+    train_gen = ExampleGenerator2D((32, 32), (0, 0), (1, 1), method='equally-spaced')
     solution_neural_net_laplace, _ = solve2D(
-        pde=laplace, condition=bc, xy_min=[0, 0], xy_max=[1, 1],
+        pde=laplace, condition=bc, xy_min=(0, 0), xy_max=(1, 1),
         net=net, max_epochs=3, train_generator=train_gen, batch_size=64
     )
-    train_gen = ExampleGenerator2D([32, 32], [0, 0], [1, 1], method='equally-spaced-noisy')
+    train_gen = ExampleGenerator2D((32, 32), (0, 0), (1, 1), method='equally-spaced-noisy')
     solution_neural_net_laplace, _ = solve2D(
-        pde=laplace, condition=bc, xy_min=[0, 0], xy_max=[1, 1],
+        pde=laplace, condition=bc, xy_min=(0, 0), xy_max=(1, 1),
         net=net, max_epochs=3, train_generator=train_gen, batch_size=64
     )
 
     with raises(ValueError):
-        train_gen = ExampleGenerator2D([32, 32], [0, 0], [1, 1], method='magic')
+        train_gen = ExampleGenerator2D((32, 32), (0, 0), (1, 1), method='magic')
     print('ExampleGenerator test passed.')
+
+    valid_gen = ExampleGenerator2D((32, 32), (0, 0), (1, 1), method='equally-spaced-noisy')
[…]
