### 2020-01-23 11b01af
for pde systems, allow user to use a single neural network
