+++ tests/test_pde.py
@@ -5,12 +5,13 @@ matplotlib.use('Agg') # use a non-GUI backend, so plots are not shown during tes
 
 from neurodiffeq import diff
 from neurodiffeq.networks import FCNN
-from neurodiffeq.pde import DirichletBVP2D, IBVP1D
-from neurodiffeq.pde import solve2D, ExampleGenerator2D, Monitor2D, make_animation
+from neurodiffeq.pde import DirichletBVP2D, IBVP1D, Condition, _network_output_2input
+from neurodiffeq.pde import solve2D, solve2D_system, ExampleGenerator2D, Monitor2D, make_animation
 
 from pytest import raises
 
 import torch
+import torch.nn as nn
 torch.manual_seed(42)
 np.random.seed(42)
 
@@ -220,7 +221,7 @@ def test_neumann_boundaries():
     )
 
     net = FCNN(n_input_units=2, n_hidden_units=32, n_hidden_layers=1)
-    
+
     solution_neural_net_heat, _ = solve2D(
         pde=heat, condition=ibvp, xy_min=(0, 0), xy_max=(L, T),
         net=net, max_epochs=300,
@@ -236,3 +237,132 @@ def test_neumann_boundaries():
     sol_net = solution_neural_net_heat(xx, tt, as_type='np')
     assert isclose(sol_net, sol_ana, atol=0.1).all()
     print('Neumann on the left Dirichlet on the right test passed.')
+
+
+def test_pde_system():
+    def _network_output_2input(net, xs, ys, ith_unit):
+        xys = torch.cat((xs, ys), 1)
+        nn_output = net(xys)
+        if ith_unit is not None:
+            return nn_output[:, ith_unit].reshape(-1, 1)
+        else:
+            return nn_output
+
+    class BCOnU(Condition):
+        """for u(x, y), impose u(x, -1) = u(x, 1) = 0; dudx(0, y) = dudy(L, y) = 0"""
+
+        def __init__(self, x_min, x_max, y_min, y_max):
+            super().__init__()
+            self.x_min = x_min
+            self.x_max = x_max
+            self.y_min = y_min
+            self.y_max = y_max
+
+        def enforce(self, net, x, y):
+            uxy = _network_output_2input(net, x, y, self.ith_unit)
+
+            x_ones = torch.ones_like(x, requires_grad=True)
+            x_ones_min = self.x_min * x_ones
+            x_ones_max = self.x_max * x_ones
+            uxminy = _network_output_2input(net, x_ones_min, y, self.ith_unit)
+            uxmaxy = _network_output_2input(net, x_ones_max, y, self.ith_unit)
+
+            x_tilde = (x - self.x_min) / (self.x_max - self.x_min)
+            y_tilde = (y - self.y_min) / (self.y_max - self.y_min)
+
+            return y_tilde * (1 - y_tilde) * (
+                    uxy - x_tilde * (self.x_max - self.x_min) * diff(uxminy, x_ones_min) \
+                    + 0.5 * x_tilde ** 2 * (self.x_max - self.x_min) * (
+                            diff(uxminy, x_ones_min) - diff(uxmaxy, x_ones_max)
+                    )
+            )
+
+    class BCOnP(Condition):
+        """for p(x, y), impose p(0, y) = p_max; p(L, y) = p_min"""
+
+        def __init__(self, x_min, x_max, p_x_min, p_x_max):
+            super().__init__()
+            self.x_min = x_min
+            self.x_max = x_max
[…]
