+++ neurodiffeq/networks.py
@@ -6,6 +6,8 @@ class FCNN(nn.Module):
 
     :param n_input_units: number of units in the input layer, defaults to 1.
     :type n_input_units: int
+    :param n_input_units: number of units in the output layer, defaults to 1.
+    :type n_input_units: int
     :param n_hidden_units: number of hidden units in each hidden layer, defaults to 32.
     :type n_hidden_units: int
     :param n_hidden_layers: number of hidden layers, defaults to 1.
@@ -13,7 +15,7 @@ class FCNN(nn.Module):
     :param actv: the activation layer used in each hidden layer, defaults to `torch.nn.Tanh`.
     :type actv: `torch.nn.Module`
     """
-    def __init__(self, n_input_units=1, n_hidden_units=32, n_hidden_layers=1, 
+    def __init__(self, n_input_units=1, n_output_units=1, n_hidden_units=32, n_hidden_layers=1,
                  actv=nn.Tanh):
         r"""Initializer method.
         """
@@ -25,7 +27,7 @@ class FCNN(nn.Module):
         for i in range(n_hidden_layers):
             layers.append(nn.Linear(n_hidden_units, n_hidden_units))
             layers.append(actv())
-        layers.append(nn.Linear(n_hidden_units, 1))
+        layers.append(nn.Linear(n_hidden_units, n_output_units))
         self.NN = torch.nn.Sequential(*layers)
 
     def forward(self, t):
