+++ neurodiffeq/pde.py
@@ -11,19 +11,45 @@ from .networks import FCNN
 from .neurodiffeq import diff
 from copy import deepcopy
 
-
-def _nn_output_2input(net, xs, ys):
+def _network_output_2input(net, xs, ys, ith_unit):
     xys = torch.cat((xs, ys), 1)
-    return net(xys)
+    nn_output = net(xys)
+    if ith_unit is not None:
+        return nn_output[:, ith_unit].reshape(-1, 1)
+    else:
+        return nn_output
+
+def _trial_solution_2input(single_net, nets, xs, ys, conditions):
+    if single_net:  # using a single net
+        us = [
+            con.enforce(single_net, xs, ys)
+            for con in conditions
+        ]
+    else:  # using multiple nets
+        us = [
+            con.enforce(net, xs, ys)
+            for con, net in zip(conditions, nets)
+        ]
+    return us
+
+class Condition:
+
+    def __init__(self):
+        self.ith_unit = None
+
+    def set_impose_on(self, ith_unit):
+        self.ith_unit = ith_unit
+
+class NoCondition2D(Condition):
+
+    def __init__(self):
+        super().__init__()
+
+    def enforce(self, net, x, y):
+        return _network_output_2input(net, x, y, self.ith_unit)
 
 
-class NoCondition2D:
-    @staticmethod
-    def enforce(net, x, y):
-        return _nn_output_2input(net, x, y)
-
-
-class DirichletBVP2D:
+class DirichletBVP2D(Condition):
     """An Dirichlet boundary value problem on a 2-D orthogonal box where :math:`x\\in[x_0, x_1]` and :math:`y\\in[y_0, y_1]`
         We are solving :math:`u(x, t)` given:
         :math:`u(x, y)\\bigg|_{x = x_0} = f_0(y)`;
@@ -52,6 +78,7 @@ class DirichletBVP2D:
     def __init__(self, x_min, x_min_val, x_max, x_max_val, y_min, y_min_val, y_max, y_max_val):
         """Initializer method
         """
+        super().__init__()
         self.x_min, self.x_min_val = x_min, x_min_val
         self.x_max, self.x_max_val = x_max, x_max_val
         self.y_min, self.y_min_val = y_min, y_min_val
@@ -72,7 +99,7 @@ class DirichletBVP2D:
             .. note::
                 `enforce` is meant to be called by the function `solve2D`.
         """
-        u = _nn_output_2input(net, x, y)
+        u = _network_output_2input(net, x, y, self.ith_unit)
         x_tilde = (x-self.x_min) / (self.x_max-self.x_min)
         y_tilde = (y-self.y_min) / (self.y_max-self.y_min)
         Axy = (1-x_tilde)*self.x_min_val(y) + x_tilde*self.x_max_val(y) + \
@@ -83,7 +110,7 @@ class DirichletBVP2D:
         return Axy + x_tilde*(1-x_tilde)*y_tilde*(1-y_tilde)*u
 
 
-class IBVP1D:
+class IBVP1D(Condition):
     """An initial boundary value problem on a 1-D range where :math:`x\\in[x_0, x_1]` and time starts at :math:`t_0`
             We are solving :math:`u(x, t)` given:
             :math:`u(x, t)\\bigg|_{t = t_0} = u_0(x)`;
@@ -120,6 +147,7 @@ class IBVP1D:
         .. note::
             A instance method `enforce` is dynamically created to enforce initial and boundary conditions. It will be called by the function `solve2D`.
         """
+        super().__init__()
         self.x_min, self.x_min_val, self.x_min_prime = x_min, x_min_val, x_min_prime
[…]
