### 2020-01-25 945efb7
add base class for ivp and bvp
