### 2020-01-25 1c85e5d
optional to use a single network for ode systems
