+++ neurodiffeq/ode.py
@@ -10,6 +10,28 @@ from .networks import FCNN
 from copy import deepcopy
 
 
+def _network_output(net, ts, ith_unit):
+    nn_output = net(ts)
+    if ith_unit is not None:
+        return nn_output[:, ith_unit].reshape(-1, 1)
+    else:
+        return nn_output
+
+
+def _trial_solution(single_net, nets, ts, conditions):
+    if single_net:  # using a single net
+        us = [
+            con.enforce(single_net, ts)
+            for con in conditions
+        ]
+    else:  # using multiple nets
+        us = [
+            con.enforce(net, ts)
+            for con, net in zip(conditions, nets)
+        ]
+    return us
+
+
 class Condition:
 
     def __init__(self):
@@ -32,7 +54,7 @@ class NoCondition(Condition):
         .. note::
             `enforce` is meant to be called by the function `solve` and `solve_system`.
         """
-        return net(t)
+        return _network_output(net, t, self.ith_unit)
 
 
 class IVP(Condition):
@@ -66,7 +88,7 @@ class IVP(Condition):
         .. note::
             `enforce` is meant to be called by the function `solve` and `solve_system`.
         """
-        x = net(t)
+        x = _network_output(net, t, self.ith_unit)
         if self.x_0_prime:
             return self.x_0 + (t-self.t_0)*self.x_0_prime + ( (1-torch.exp(-t+self.t_0))**2 )*x
         else:
@@ -106,7 +128,7 @@ class DirichletBVP(Condition):
         .. note::
             `enforce` is meant to be called by the function `solve` and `solve_system`.
         """
-        x = net(t)
+        x = _network_output(net, t, self.ith_unit)
         t_tilde = (t-self.t_0) / (self.t_1-self.t_0)
         return self.x_0*(1-t_tilde) + self.x_1*t_tilde + (1-torch.exp((1-t_tilde)*t_tilde))*x
 
@@ -187,7 +209,7 @@ class Monitor:
         # input for neural network
         self.ts_ann = torch.linspace(t_min, t_max, 100, requires_grad=True).reshape((-1, 1))
 
-    def check(self, nets, conditions, loss_history):
+    def check(self, single_net, nets, conditions, loss_history):
         r"""Draw 2 plots: One shows the shape of the current solution. The other shows the history training loss and validation loss.
 
         :param nets: The neural networks that approximates the ODE (system).
@@ -200,16 +222,12 @@ class Monitor:
         .. note::
             `check` is meant to be called by the function `solve` and `solve_system`.
         """
-        n_dependent = len(conditions)
-
-        vs = [
-            con.enforce(net, self.ts_ann).detach().cpu().numpy()
-            for con, net in zip(conditions, nets)
-        ]
+        us = _trial_solution(single_net, nets, self.ts_ann, conditions)
+        us = [u.detach().cpu().numpy() for u in us]
 
         self.ax1.clear()
-        for i in range(n_dependent):
-            self.ax1.plot(self.ts_plt, vs[i], label=f'variable {i}')
+        for i, u in enumerate(us):
+            self.ax1.plot(self.ts_plt, u, label=f'variable {i}')
         self.ax1.legend()
         self.ax1.set_title('solutions')
 
@@ -228,7 +246,7 @@ class Monitor:
 
 
 def solve(
-        ode, condition, t_min, t_max,
+        ode, condition, t_min=None, t_max=None,
         net=None, train_generator=None, shuffle=True, valid_generator=None,
         optimizer=None, criterion=None, batch_size=16,
         max_epochs=1000,
@@ -244,9 +262,9 @@ def solve(
[…]
