### 2020-01-26 1306148
additional loss term for ode
