### 2020-01-28 2c8979c
custom metrics other than loss + now training loss is calculated after each epoch instead of using an average of the loss of each batch
