+++ neurodiffeq/pde.py
@@ -319,7 +319,7 @@ class Monitor2D:
         self.xs_ann, self.ys_ann = xs_ann.reshape(-1, 1), ys_ann.reshape(-1, 1)
 
 
-    def check(self, single_net, nets, conditions, loss_history):
+    def check(self, single_net, nets, conditions, history):
         r"""Draw 2 plots: One shows the shape of the current solution (with heat map). The other shows the history training loss and validation loss.
 
         :param single_net: The neural network that approximates the PDE.
@@ -328,8 +328,8 @@ class Monitor2D:
         :type nets: list [`torch.nn.Module`]
         :param conditions: The initial/boundary condition of the PDE.
         :type conditions: list [`neurodiffeq.pde.DirichletBVP2D` or `neurodiffeq.pde.IBVP1D` or `neurodiffeq.pde.NoCondition`]
-        :param loss_history: The history of training loss and validation loss. The 'train' entry is a list of training loss and 'valid' entry is a list of validation loss.
-        :type loss_history: dict['train': list[float], 'valid': list[float]]
+        :param history: The history of training loss and validation loss. The 'train' entry is a list of training loss and 'valid' entry is a list of validation loss.
+        :type history: dict['train': list[float], 'valid': list[float]]
 
         .. note::
             `check` is meant to be called by the function `solve2D`.
@@ -338,7 +338,7 @@ class Monitor2D:
         if not self.fig:
             # initialize the figure and axes here so that the Monitor knows the number of dependent variables and
             # size of the figure, number of the subplots, etc.
-            n_axs = len(conditions)+1  # one for each dependent variable, another one for training and validation loss
+            n_axs = len(conditions)+2  # one for each dependent variable, plus one for training and validation loss, plus one for metrics
             n_row, n_col = (n_axs+1) // 2, 2
             self.fig = plt.figure(figsize=(20, 8*n_row))
             for i in range(n_axs):
@@ -358,11 +358,22 @@ class Monitor2D:
             self.cbs[i] = self.fig.colorbar(cax, ax=ax)
             ax.set_title(f'u[{i}](x, y)')
 
+        self.axs[-2].clear()
+        self.axs[-2].plot(history['train_loss'], label='training loss')
+        self.axs[-2].plot(history['valid_loss'], label='validation loss')
+        self.axs[-2].set_title('loss during training')
+        self.axs[-2].set_ylabel('loss')
+        self.axs[-2].set_xlabel('epochs')
+        self.axs[-2].set_yscale('log')
+        self.axs[-2].legend()
+
         self.axs[-1].clear()
-        self.axs[-1].plot(loss_history['train'], label='training loss')
-        self.axs[-1].plot(loss_history['valid'], label='validation loss')
-        self.axs[-1].set_title('loss during training')
-        self.axs[-1].set_ylabel('loss')
+        for metric_name, metric_values in history.items():
+            if metric_name == 'train_loss' or metric_name == 'valid_loss':
+                continue
+            self.axs[-1].plot(metric_values, label=metric_name)
+        self.axs[-1].set_title('metrics during training')
+        self.axs[-1].set_ylabel('metrics')
         self.axs[-1].set_xlabel('epochs')
         self.axs[-1].set_yscale('log')
         self.axs[-1].legend()
@@ -374,7 +385,7 @@ class Monitor2D:
 
 def solve2D(
         pde, condition, xy_min=None, xy_max=None,
-        net=None, train_generator=None, shuffle=True, valid_generator=None, optimizer=None, criterion=None, additional_loss_term=None,
+        net=None, train_generator=None, shuffle=True, valid_generator=None, optimizer=None, criterion=None, additional_loss_term=None, metrics=None,
         batch_size=16,
         max_epochs=1000,
         monitor=None, return_internal=False, return_best=False
[…]
