### 2020-01-28 1998d8e
custom metrics other than loss + now training loss is calculated after each epoch instead of using an average of the loss of each batch (for ode module)
