+++ neurodiffeq/ode.py
@@ -201,23 +201,24 @@ class Monitor:
         """
         self.using_non_gui_backend = matplotlib.get_backend() is 'agg'
         self.check_every = check_every
-        self.fig = plt.figure(figsize=(20, 8))
-        self.ax1 = self.fig.add_subplot(121)
-        self.ax2 = self.fig.add_subplot(122)
+        self.fig = plt.figure(figsize=(30, 8))
+        self.ax1 = self.fig.add_subplot(131)
+        self.ax2 = self.fig.add_subplot(132)
+        self.ax3 = self.fig.add_subplot(133)
         # input for plotting
         self.ts_plt = np.linspace(t_min, t_max, 100)
         # input for neural network
         self.ts_ann = torch.linspace(t_min, t_max, 100, requires_grad=True).reshape((-1, 1))
 
-    def check(self, single_net, nets, conditions, loss_history):
+    def check(self, single_net, nets, conditions, history):
         r"""Draw 2 plots: One shows the shape of the current solution. The other shows the history training loss and validation loss.
 
         :param nets: The neural networks that approximates the ODE (system).
         :type nets: list[`torch.nn.Module`]
         :param conditions: The initial/boundary conditions of the ODE (system).
         :type conditions: list[`neurodiffeq.ode.IVP` or `neurodiffeq.ode.DirichletBVP` or `neurodiffeq.ode.NoCondition`]
-        :param loss_history: The history of training loss and validation loss. The 'train' entry is a list of training loss and 'valid' entry is a list of validation loss.
-        :type loss_history: dict['train': list[float], 'valid': list[float]]
+        :param history: The history of training loss and validation loss. The 'train' entry is a list of training loss and 'valid' entry is a list of validation loss.
+        :type history: dict['train': list[float], 'valid': list[float]]
 
         .. note::
             `check` is meant to be called by the function `solve` and `solve_system`.
@@ -232,13 +233,24 @@ class Monitor:
         self.ax1.set_title('solutions')
 
         self.ax2.clear()
-        self.ax2.plot(loss_history['train'], label='training loss')
-        self.ax2.plot(loss_history['valid'], label='validation loss')
+        self.ax2.plot(history['train_loss'], label='training loss')
+        self.ax2.plot(history['valid_loss'], label='validation loss')
         self.ax2.set_title('loss during training')
         self.ax2.set_ylabel('loss')
         self.ax2.set_xlabel('epochs')
         self.ax2.set_yscale('log')
         self.ax2.legend()
+
+        self.ax3.clear()
+        for metric_name, metric_values in history.items():
+            if metric_name == 'train_loss' or metric_name == 'valid_loss':
+                continue
+            self.ax3.plot(metric_values, label=metric_name)
+        self.ax3.set_title('metrics during training')
+        self.ax3.set_ylabel('metrics')
+        self.ax3.set_xlabel('epochs')
+        self.ax3.set_yscale('log')
+        self.ax3.legend()
 
         self.fig.canvas.draw()
         if not self.using_non_gui_backend:
@@ -248,7 +260,7 @@ class Monitor:
 def solve(
         ode, condition, t_min=None, t_max=None,
         net=None, train_generator=None, shuffle=True, valid_generator=None,
-        optimizer=None, criterion=None, additional_loss_term=None, batch_size=16,
+        optimizer=None, criterion=None, additional_loss_term=None, metrics=None, batch_size=16,
         max_epochs=1000,
         monitor=None, return_internal=False,
         return_best=False
@@ -278,6 +290,9 @@ def solve(
     :type criterion: `torch.nn.modules.loss._Loss`, optional
     :param additional_loss_term: Extra terms to add to the loss function besides the part specified by `criterion`. The input of `additional_loss_term` should be the same as `ode`
     :type additional_loss_term: function
[…]
