### 2020-01-28 7e29ef0
when metrics not specified default to an empty dict
