+++ neurodiffeq/pde.py
@@ -579,6 +579,8 @@ def solve2D_system(
         optimizer = optim.Adam(all_parameters, lr=0.001)
     if not criterion:
         criterion = nn.MSELoss()
+    if metrics is None:
+        metrics = {}
 
     history = {}
     history['train_loss'] = []
