### 2020-01-29 9b15e2f
break down a big test into smaller ones so that travisCI wont kill this build because of the lack of response + corrected mse formula
