### 2020-01-29 1bed7cb
break down a big test into smaller ones so that travisCI wont kill this build because of the lack of response + corrected mse formula
