+++ tests/test_pde.py
@@ -254,117 +254,117 @@ def test_neumann_boundaries_3():
     print('Neumann on the left Dirichlet on the right test passed.')
 
 
-def test_pde_system():
-    def _network_output_2input(net, xs, ys, ith_unit):
-        xys = torch.cat((xs, ys), 1)
-        nn_output = net(xys)
-        if ith_unit is not None:
-            return nn_output[:, ith_unit].reshape(-1, 1)
-        else:
-            return nn_output
-
-    class BCOnU(Condition):
-        """for u(x, y), impose u(x, -1) = u(x, 1) = 0; dudx(0, y) = dudy(L, y) = 0"""
-
-        def __init__(self, x_min, x_max, y_min, y_max):
-            super().__init__()
-            self.x_min = x_min
-            self.x_max = x_max
-            self.y_min = y_min
-            self.y_max = y_max
-
-        def enforce(self, net, x, y):
-            uxy = _network_output_2input(net, x, y, self.ith_unit)
-
-            x_ones = torch.ones_like(x, requires_grad=True)
-            x_ones_min = self.x_min * x_ones
-            x_ones_max = self.x_max * x_ones
-            uxminy = _network_output_2input(net, x_ones_min, y, self.ith_unit)
-            uxmaxy = _network_output_2input(net, x_ones_max, y, self.ith_unit)
-
-            x_tilde = (x - self.x_min) / (self.x_max - self.x_min)
-            y_tilde = (y - self.y_min) / (self.y_max - self.y_min)
-
-            return y_tilde * (1 - y_tilde) * (
-                    uxy - x_tilde * (self.x_max - self.x_min) * diff(uxminy, x_ones_min) \
-                    + 0.5 * x_tilde ** 2 * (self.x_max - self.x_min) * (
-                            diff(uxminy, x_ones_min) - diff(uxmaxy, x_ones_max)
-                    )
-            )
-
-    class BCOnP(Condition):
-        """for p(x, y), impose p(0, y) = p_max; p(L, y) = p_min"""
-
-        def __init__(self, x_min, x_max, p_x_min, p_x_max):
-            super().__init__()
-            self.x_min = x_min
-            self.x_max = x_max
-            self.p_x_min = p_x_min
-            self.p_x_max = p_x_max
-
-        def enforce(self, net, x, y):
-            uxy = _network_output_2input(net, x, y, self.ith_unit)
-            x_tilde = (x - self.x_min) / (self.x_max - self.x_min)
-
-            return (1 - x_tilde) * self.p_x_min + x_tilde * self.p_x_max \
-                   + x_tilde * (1 - x_tilde) * uxy
-
-    L = 2.0
-    mu = 1.0
-    P1, P2 = 1.0, 0.0
-    def poiseuille(u, v, p, x, y):
-        return [
-            mu * (diff(u, x, order=2) + diff(u, y, order=2)) - diff(p, x),
-            mu * (diff(v, x, order=2) + diff(v, y, order=2)) - diff(p, y),
-            diff(u, x) + diff(v, y)
-        ]
-    def zero_divergence(u, v, p, x, y):
-        return torch.sum( (diff(u, x) + diff(v, y))**2 )
-
-    bc_on_u = BCOnU(
-        x_min=0,
-        x_max=L,
-        y_min=-1,
-        y_max=1,
-    )
-    bc_on_v = DirichletBVP2D(
-        x_min=0, x_min_val=lambda y: 0,
-        x_max=L, x_max_val=lambda y: 0,
[…]
