+++ neurodiffeq/pde.py
@@ -32,6 +32,7 @@ def _trial_solution_2input(single_net, nets, xs, ys, conditions):
         ]
     return us
 
+
 class Condition:
 
     def __init__(self):
@@ -39,6 +40,9 @@ class Condition:
 
     def set_impose_on(self, ith_unit):
         self.ith_unit = ith_unit
+
+    def enforce(self, net, *dimensions):
+        raise NotImplementedError
 
 class NoCondition2D(Condition):
 
@@ -292,6 +296,18 @@ class ExampleGenerator2D:
             )
         else:
             raise ValueError(f'Unknown method: {method}')
+
+
+class PredefinedExampleGenerator2D:
+
+    def __init__(self, xs, ys):
+        self.size = len(xs)
+        x = torch.tensor(xs, requires_grad=True, dtype=torch.float32)
+        y = torch.tensor(ys, requires_grad=True, dtype=torch.float32)
+        self.x, self.y = x.flatten(), y.flatten()
+
+    def get_examples(self):
+        return self.x, self.y
 
 
 class Monitor2D:
@@ -712,3 +728,354 @@ def make_animation(solution, xs, ts):
     return animation.FuncAnimation(
         fig, run, u_gen, blit=True, interval=50, repeat=False
     )
+
+############################# arbitraty boundary conditions #############################
+
+# CONSTANTS
+ROUND_TO_ZERO = 1e-7
+K = 5.0
+ALPHA = 5.0
+
+
+class Point:
+
+    def __repr__(self):
+        return f'Point({self.loc})'
+
+    def __init__(self, loc):
+        self.loc = tuple(float(d) for d in loc)
+        self.dim = len(loc)
+
+
+class DirichletControlPoint(Point):
+
+    def __repr__(self):
+        return f'DirichletControlPoint({self.loc}, val={self.val})'
+
+    def __init__(self, loc, val):
+        super().__init__(loc)
+        self.val = float(val)
+
+
+class NeumannControlPoint(Point):
+
+    def __repr__(self):
+        return f'NeumannControlPoint({self.loc}, val={self.val}, ' + \
+               f'normal_vector={self.normal_vector})'
+
+    def __init__(self, loc, val, normal_vector):
+        super().__init__(loc)
+        self.val = float(val)
+        scale = sum(d ** 2 for d in normal_vector) ** 0.5
+        self.normal_vector = tuple(d / scale for d in normal_vector)
+
+
+class Condition:
+
+    def __init__(self):
+        self.ith_unit = None
+
+    def set_impose_on(self, ith_unit):
+        self.ith_unit = ith_unit
+
+    def enforce(self, net, *dimensions):
+        raise NotImplementedError
+
+
+class CustomBoundaryCondition(Condition):
+
+    def __init__(self, dirichlet_control_points, neumann_control_points, center_point):
+        super().__init__()
+
+        # for Dirichlet control points, drop deplicates and sort 'clockwise'
+        self.dirichlet_control_points = self._clean_control_points(dirichlet_control_points, center_point)
+        # fit Neumann boundary condition (A_D(x) in MacFall's paper)
+        self.a_d_interp = InterpolatorCreator.fit_surface(self.dirichlet_control_points)
+        # fit Dirichlet length factor (L_D(x) in MacFall's paper)
+        self.l_d_interp = InterpolatorCreator.fit_length_factor(self.dirichlet_control_points)
+
+        # for Naumann control points, drop deplicates and sort 'clockwise'
+        self.neumann_control_points = self._clean_control_points(neumann_control_points, center_point)
+        # fit Neumann boundary condition (g(x) in MacFall's paper)
+        self.g_interp = InterpolatorCreator.fit_surface(self.neumann_control_points)
+        # fit Naumann length factor (L_M(x) in MacFall's paper)
[…]
