### 2020-02-04 #1931 @dawbarton
Overall, the paper reads well and covers the material well. I agree with the comments of the other reviewer. It would be good to mention the dependence on torch - I installed the respository directly rather than via conda (since I don't have conda installed) and so also needed to install torch separately. (Easy enough via pip, but still needs doing.)

The examples run well on my laptop, though I wouldn't necessarily describe them as real-world examples. (I'd be curious to see it applied to a truly real-world example where there are significant nonlinearities.)

The justification for this package in the paper is perfectly reasonable, though the justification in the docs is far less accurate. The line "While these methods are effective and adequate, their solutions are discrete." is not entirely correct; there are many numerical methods for solving DEs that produce a continuous output.

The methods section of the paper is very broad-brush in that it doesn't go into any detail about the formulation of the problem and only shows a very high-level interpretation of the differential equation. There are many other questions I'm interested in, such as how is the derivative operator discretized, though I appreciate that this is a software paper rather than a paper on the methods and so I can understand why it isn't included.

All the checkboxes are OK. From my perpective, other than mentioning about torch in the installation section it's good to go.
