### 2020-02-05 12e6e3e
fix random seed for pde_spherical tests
