### 2020-02-05 ef19434
do not try to plot legends when there is no custom metrics
