+++ neurodiffeq/ode.py
@@ -250,7 +250,9 @@ class Monitor:
         self.ax3.set_ylabel('metrics')
         self.ax3.set_xlabel('epochs')
         self.ax3.set_yscale('log')
-        self.ax3.legend()
+        # if there's not custom metrics, then there won't be any labels in this axis
+        if len(history) > 2:
+            self.axs[-1].legend()
 
         self.fig.canvas.draw()
         if not self.using_non_gui_backend:
