### 2020-02-06 b5933ed
custom valid_gen for monitor
