### 2020-02-10 0733c10
when user are not specifying neumann control points, assume there is no neumann boundary condition
