### 2020-02-11 fbd97c2
add instruction to use venv
