+++ paper/paper.md
@@ -15,20 +15,23 @@ authors:
   - name: Marios Mattheakis
     affiliation: 1
   - name: Shuheng Liu
+    affiliation: 2
   - name: Devansh Agarwal
-    affiliation: 2, 3
+    affiliation: 3, 4
   - name: Marco Di Giovanni
-    affiliation: 4
+    affiliation: 5
 
 affiliations:
  - name: Institute for Applied Computational Science, Harvard University, Cambridge, MA, United States
    index: 1
- - name: Department of Physics and Astronomy, Virginia University, Morgantown, WV, United States
+ - name: Chongqing University
    index: 2
- - name: Center for Gravitational Waves and Cosmology, West Virginia University, Morgantown, WV, USA
+ - name: Department of Physics and Astronomy, Virginia University, Morgantown, WV, United States
    index: 3
- - name: Politecnico di Milano, Milano, Italy
+ - name: Center for Gravitational Waves and Cosmology, West Virginia University, Morgantown, WV, USA
    index: 4
+ - name: Politecnico di Milano, Milano, Italy
+   index: 5
 date: 31 October 2019
 bibliography: paper.bib
 ---
