### 2020-02-12 #1931 @feiyu-chen96
Thank you @marcharper and @dawbarton for your time and reviews!

## Review 1 (@marcharper)

1. on the authors: There are 3 new contributors to the package since the initial submission. They've be added to the authors (Shuheng @Johnny-Wish, Devansh @devanshkv, and Marco @marco-digio).
2. on installation and testing: Thanks! The installation section is now updated to include commands to install dependencies. Since those are still extra steps required from the user that could have been done by us, we will publish the package on PyPI in the future. The testing command is now included as well.

## Review 2 (@dawbarton)

1. on installation: Thanks! For the virtual environment, we've added `venv` as an alternative to `conda`. It comes with Python standard library so the user won't need to install extra packages. The requirements.txt now includes torch as well.
1. on real-world examples: Yes we agree that the examples included in the test and documentation are more or less toy problems. We are working on some more real-world problems at present. There's an ongoing project solving equations in general relativity. I'll also include some fluid problems in documents as my thesis progresses.
3. on documentation: Thanks! The motivation section in the documentation and README is indeed not accurate. It is now updated. For the methods section, we have added more explanation on how the software is implemented.
