### 2020-02-12 c778037
error goes down to e-14 when we use float64
