### 2020-02-19 #1931 @drvinceknight
@openjournals/joss-eics this paper should be ready for publication now :+1:
