### 2020-02-19 4b71d2d
convert training loss Variable class to its numeric value to fix memory issue
