### 2020-02-19 503069d
Update pde.py
convert training loss Variable class to its numeric value to fix memory issue
