### 2020-02-19 #1931 @kyleniemeyer
Hi @feiyu-chen96, regarding the article, I have just two minor corrections:
- in the first paragraph, you should combine the references to Seefeldt and Smith, like this: `[@author1:2001; @author2:2001]`, which will make them look like "(Author1 et al., 2001; Author2 et al., 2002)"
- in the second paragraph, insert "and" after "(2) good interpolation properties, "
