### 2020-02-19 #1931 @feiyu-chen96
@kyleniemeyer Thanks! I've updated the paper.
