### 2020-02-19 #1931 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
